requirements.txt
streamlit==1.34.0
python-dotenv==1.0.1
langchain==0.2.2
langchain-community==0.2.3
langchain-anthropic==0.1.15
langchain-voyageai==0.1.1
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.74
anthropic==0.28.0
voyageai==0.2.3
openai==1.31.1
faiss-cpu==1.8.0
tqdm==4.66.4
bs4==0.0.2
chardet==5.2.0
requests==2.32.2
better_profanity==0.7.0
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0